requirements.txt
torch
yolov5
opencv-python
numpy
Pillow
customtkinter
tk
